Overhaul of disk handling
Previously, a file handle was kept open for each entry in the backing
store. This could run up against user/system limits on open handles for
a large set of backing files, and also wasted memory keeping open unused
write/read buffers.

The new design provides functions to spawn read/write handles. These are
created for a read/write, and then dropped at the end of the operation.
The extra time to call a function dynamically is minor when execution
time is dominated by I/O.

Docs are out of date with this commit, to be resolved later.