Refactor container handling
There is now support for containers that do not directly dereference
into a slice. The slice methods were replaced with c_* calls and some
iterator overhaul.

BorrowExtenders now use a box pin to avoid moves and guarantee that the
borrows are preserved. This does cost some heap allocation, but it is
the simplest way to prevent moving parent reference shenanigans.